requirements.txt
langchain
langchain-text-splitters
langchain-community
langchain-chroma
langchain-openai
chromadb
openai
python-dotenv
markdown
tiktoken
